41 kopis 타입 변환 로직 어댑터로 분리 (#52)
* refactor: KOPIS 타입 변환 로직 kopis.adapter.ts로 분리 (#41)

* refactor: boxOfficeToPerformance 어댑터 적용

* ♻️[Refactor] KOPIS raw 타입 및 파서를 adapter 레이어로 분리

- src/api/adapters/types/kopis.ts 신규 생성: KopisRaw* 타입 이동
- src/api/adapters/kopis.parser.ts 신규 생성: parseKopisXml, toArray, parse* 이동
- src/api/adapters/kopis.adapter.ts: raw 타입 import 경로 변경
- src/shared/types/performance.ts: Raw 타입 제거, UI 타입만 유지
- src/shared/utils/kopis.ts: date 유틸(kopisDateToFormDate, resolveShowDate)만 남김
- pages/api/kopis/*.ts: parser import 경로 → @/api/adapters/kopis.parser

* ♻️[Refactor] kopisDateToFormDate를 date.ts로 통합, kopis.ts 제거

- shared/utils/date.ts: kopisDateToFormDate 추가 (KOPIS 섹션)
- useRecordEditorForm.ts: resolveShowDate 제거 → kopisDateToFormDate(p.startDate) 직접 호출
- shared/utils/kopis.ts: 삭제 (역할 없어짐)

* 🔥[Remove] 미사용 파일 삭제
