--- src/shared/hooks/useIsMobile.ts
@@ -1,12 +1,0 @@
-import { useEffect, useState } from "react";
-
-export function useIsMobile(breakpoint = 1024) {
-  const [isMobile, setIsMobile] = useState(true);
-  useEffect(() => {
-    const check = () => setIsMobile(window.innerWidth < breakpoint);
-    check();
-    window.addEventListener("resize", check);
-    return () => window.removeEventListener("resize", check);
-  }, [breakpoint]);
-  return isMobile;
-}
